requirements.txt
confluent-kafka==1.8.2
flask==2.2.2
kafka-python
